reporter/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
google-analytics-data>=0.18.0
google-auth>=2.27.0
